+++ src/scenes.js
@@ -54,7 +54,7 @@ ScriptsMorph*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.scenes = '2026-March-10';
+modules.scenes = '2026-April-15';
 
 // Projecct /////////////////////////////////////////////////////////
 
@@ -63,7 +63,7 @@ modules.scenes = '2026-March-10';
 
 // Project instance creation:
 
-function Project(scenes, current) {
+function Project(scenes, current, temporary = false) {
     var projectScene;
 
     this.scenes = scenes || new List();
@@ -88,6 +88,9 @@ function Project(scenes, current) {
 
     // for undeleting scenes - do not persist
     this.trash = [];
+
+    // for refreshing the ide
+    this.isTemporary = temporary;
 }
 
 Project.prototype.initialize = function () {
@@ -125,8 +128,20 @@ function Scene(aStageMorph) {
     this.showPaletteButtons = true;
     this.role = null; // null (default), "template" or "tutorial"
     this.createdFromTemplate = false;
-    this.template = null; // {name: str, version: str, hide: nested list}
     this.hideSprites = false;
+
+    // template settings
+    this.template = {
+        name: null,
+        version: null,
+        hide: null,
+        lang: undefined,
+        zoom: undefined,
+        scale: undefined,
+        fade: undefined,
+        flat: undefined,
+        bright: undefined
+    };
 
     // cached IDE state
     this.sprites = new List();
@@ -232,6 +247,13 @@ Scene.prototype.applyGlobalSettings = function () {
     SpriteMorph.prototype.penColorModel = this.penColorModel;
     SpriteMorph.prototype.blocks = this.blocks;
     ScriptsMorph.prototype.enforceTypes = this.enforceTypes;
+};
+
+// Scene embedded template settings
+
+Scene.prototype.hasEmbeddedTemplateSettings = function () {
+    return ['lang', 'zoom', 'scale', 'fade', 'flat', 'bright'].some(any =>
+        this.template[any] !== undefined);
 };
 
 // Scene ops: